requirements.txt
fastapi==0.111.0
pydantic[email]>=2.9.0
python-dotenv>=1.0.0
